+++ fluiddyn/output/__init__.py
@@ -11,6 +11,7 @@ Scientific output toolkit
    figs
    movies
    latextables
+   colorchart
 
 """