+++ CHANGES.rst
@@ -1,11 +1,27 @@
 
-0.0.9a
+0.0.11
 ------
+
+- User configuration files.
+
+- Add color charts to choose the colors in figures.
+
+0.0.10
+------
+
+- Logger for error logging.
+
+- Utility to use comma separated values (csv) files.
+
+- Tiny package for readthedocs.
+
+0.0.9
+-----
 
 - New parameter container (API slightly changed).
 
-0.0.8a
-------
+0.0.8
+-----
 
 - The fluiddyn package now only contains base files for the FluidDyn
   project. Other packages (fluidsim, fluidlab, ...) provide other